spec setup more standard with rails_helper
