+++ spec/spec_helper.rb
@@ -1,19 +1,4 @@
-ENV['RAILS_ENV'] ||= 'test'
-
-unless ENV['CI']
-  require 'pry' # for debugging specs
-  require 'byebug'
-end
-
-Dir.glob(File.join(__dir__, 'fixtures', '**', '*.rb'), &method(:require)) # load all fixture files
-
-require 'rspec/matchers'
-require 'equivalent-xml'
 require 'simplecov'
-
-require 'webmock/rspec'
-WebMock.enable!
-
 SimpleCov.profiles.define 'sul_pub' do
   add_filter '.gems'
   add_filter '/config/environments/'
@@ -33,149 +18,100 @@ SimpleCov.profiles.define 'sul_pub' do
 end
 SimpleCov.start 'sul_pub'
 
-require File.expand_path('../../config/environment', __FILE__)
-
-ActiveRecord::Migration.maintain_test_schema!
-
-require 'rspec/rails'
-require 'factory_bot_rails'
-
+# This file was generated by the `rails generate rspec:install` command. Conventionally, all
+# specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
+# The generated `.rspec` file contains `--require spec_helper` which will cause
+# this file to always be loaded, without a need to explicitly require it in any
+# files.
+#
+# Given that it is always loaded, you are encouraged to keep this file as
+# light-weight as possible. Requiring heavyweight dependencies from this file
+# will add to the boot time of your test suite on EVERY test run, even for an
+# individual file that may not need all of that loaded. Instead, consider making
+# a separate helper file that requires the additional dependencies and performs
+# the additional setup, and require it from the spec files that actually need
+# it.
+#
+# See http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
 RSpec.configure do |config|
-  config.include RSpec::Rails::RequestExampleGroup, type: :request, file_path: %r{spec/api}
-  config.include RSpec::Rails::RequestExampleGroup, type: :request, file_path: %r{spec/lib}
+  # rspec-expectations config goes here. You can use an alternate
+  # assertion/expectation library such as wrong or the stdlib/minitest
+  # assertions if you prefer.
+  config.expect_with :rspec do |expectations|
+    # This option will default to `true` in RSpec 4. It makes the `description`
+    # and `failure_message` of custom matchers include text for helper methods
+    # defined using `chain`, e.g.:
+    #     be_bigger_than(2).and_smaller_than(4).description
+    #     # => "be bigger than 2 and smaller than 4"
+    # ...rather than:
+    #     # => "be bigger than 2"
+    expectations.include_chain_clauses_in_custom_matcher_descriptions = true
+  end
 
-  config.include FactoryBot::Syntax::Methods
+  # rspec-mocks config goes here. You can use an alternate test double
+  # library (such as bogus or mocha) by changing the `mock_with` option here.
+  config.mock_with :rspec do |mocks|
+    # Prevents you from mocking or stubbing a method that does not exist on
+    # a real object. This is generally recommended, and will default to
+    # `true` in RSpec 4.
+    # NOTE: 3 specs fail when this line is uncommented
+    # mocks.verify_partial_doubles = true
+  end
 
-  # Remove this line if you're not using ActiveRecord or ActiveRecord fixtures
-  # config.fixture_path = "#{::Rails.root}/spec/fixtures"
+  # This option will default to `:apply_to_host_groups` in RSpec 4 (and will
+  # have no way to turn it off -- the option exists only for backwards
+  # compatibility in RSpec 3). It causes shared context metadata to be
+  # inherited by the metadata hash of host groups and examples, rather than
+  # triggering implicit auto-inclusion in groups with matching metadata.
+  config.shared_context_metadata_behavior = :apply_to_host_groups
 
   # Allows RSpec to persist some state between runs in order to support
   # the `--only-failures` and `--next-failure` CLI options. We recommend
   # you configure your source control system to ignore this file.
-  config.example_status_persistence_file_path = 'spec/examples.txt'
-  
-  # If you're not using ActiveRecord, or you'd prefer not to run each of your
-  # examples within a transaction, remove the following line or assign false
-  # instead of true.
-  config.use_transactional_fixtures = true
+  config.example_status_persistence_file_path = "spec/examples.txt"
 
-  # If true, the base class of anonymous controllers will be inferred
-  # automatically. This will be the default behavior in future versions of
-  # rspec-rails.
-  config.infer_base_class_for_anonymous_controllers = false
+# The settings below are suggested to provide a good initial experience
+# with RSpec, but feel free to customize to your heart's content.
+=begin
+  # This allows you to limit a spec run to individual examples or groups
+  # you care about by tagging them with `:focus` metadata. When nothing
+  # is tagged with `:focus`, all examples get run. RSpec also provides
+  # aliases for `it`, `describe`, and `context` that include `:focus`
+  # metadata: `fit`, `fdescribe` and `fcontext`, respectively.
+  config.filter_run_when_matching :focus
+
+  # Limits the available syntax to the non-monkey patched syntax that is
+  # recommended. For more details, see:
+  #   - http://rspec.info/blog/2012/06/rspecs-new-expectation-syntax/
+  #   - http://www.teaisaweso.me/blog/2013/05/27/rspecs-new-message-expectation-syntax/
+  #   - http://rspec.info/blog/2014/05/notable-changes-in-rspec-3/#zero-monkey-patching-mode
+  config.disable_monkey_patching!
+
+  # Print the 10 slowest examples and example groups at the
+  # end of the spec run, to help surface which specs are running
+  # particularly slow.
+  config.profile_examples = 10
+=end
+
+  # Many RSpec users commonly either run the entire suite or an individual
+  # file, and it's useful to allow more verbose output when running an
+  # individual spec file.
+  if config.files_to_run.one?
+    # Use the documentation formatter for detailed output,
+    # unless a formatter has already been configured
+    # (e.g. via a command-line flag).
+    config.default_formatter = "doc"
+  end
 
   # Run specs in random order to surface order dependencies. If you find an
   # order dependency and want to debug it, you can fix the order by providing
   # the seed, which is printed after each run.
   #     --seed 1234
-  config.order = 'random'
+  config.order = :random
 
-  # rspec-rails 3 will no longer automatically infer an example group's spec type
-  # from the file location. You can explicitly opt-in to the feature using this
-  # config option.
-  # To explicitly tag specs without using automatic inference, set the `:type`
-  # metadata manually:
-  #
-  #     describe ThingsController, :type => :controller do
-  #       # Equivalent to being in spec/controllers
-  #     end
-  config.infer_spec_type_from_file_location!
+  # Seed global randomization in this process using the `--seed` CLI option.
+  # Setting this allows you to use `--seed` to deterministically reproduce
+  # test failures related to randomization by passing the same `--seed` value
+  # as the one that triggered the failure.
+  Kernel.srand config.seed
 end
-
-RSpec::Matchers.define_negated_matcher :exclude, :include
-RSpec::Matchers.define_negated_matcher :not_change, :change
-
-require 'vcr'
-VCR.configure do |c|
-  c.cassette_library_dir = 'fixtures/vcr_cassettes'
-  c.allow_http_connections_when_no_cassette = true
-  c.hook_into :webmock
-  c.default_cassette_options = {
-    :record => :new_episodes,  # :once is default
-  }
-  c.configure_rspec_metadata!
-  c.filter_sensitive_data('private_bearer_token') do |interaction|
-    auth = interaction.request.headers['Authorization']
-    if auth.is_a? Array
-      bearer = auth.select { |a| a =~ /bearer/i }.first
-      bearer.gsub(/bearer /i, '') if bearer
-    end
-  end
-  c.filter_sensitive_data('Settings.SCIENCEWIRE.HOST') { Settings.SCIENCEWIRE.HOST }
-  c.filter_sensitive_data('Settings.SCIENCEWIRE.LICENSE_ID') { Settings.SCIENCEWIRE.LICENSE_ID }
-  c.filter_sensitive_data('Settings.PUBMED.API_KEY') { Settings.PUBMED.API_KEY }
-  c.filter_sensitive_data('Settings.MAIS.BASE_URL') { Settings.MAIS.BASE_URL }
-  c.filter_sensitive_data('Settings.MAIS.CLIENT_ID') { Settings.MAIS.CLIENT_ID }
-  c.filter_sensitive_data('Settings.MAIS.CLIENT_SECRET') { Settings.MAIS.CLIENT_SECRET }
-  c.filter_sensitive_data('Settings.ORCID.CLIENT_ID') { Settings.ORCID.CLIENT_ID }
-  c.filter_sensitive_data('Settings.ORCID.CLIENT_SECRET') { Settings.ORCID.CLIENT_SECRET }
-
-  # WOS Links-AMR filters
-  (links_username, links_password) = Base64.decode64(Settings.WOS.AUTH_CODE).split(':', 2)
-  c.filter_sensitive_data('links_username') do |interaction|
-    links_username if interaction.request.uri.include? Clarivate::LinksClient::LINKS_HOST
-  end
-  c.filter_sensitive_data('links_password') do |interaction|
-    links_password if interaction.request.uri.include? Clarivate::LinksClient::LINKS_HOST
-  end
-
-  # ORCID API filters
-  c.filter_sensitive_data('refresh_token') do |interaction|
-    if interaction.request.uri.include? Settings.ORCID.BASE_AUTH_URL
-      token_match = interaction.response.body.match(/"refresh_token":"(.*?)"/)
-      token_match.captures.first if token_match
-    end
-  end
-
-  # CAP API filters
-  c.filter_sensitive_data('Settings.CAP.TOKEN_USER:Settings.CAP.TOKEN_PASS') do
-    Base64.strict_encode64("#{Settings.CAP.TOKEN_USER}:#{Settings.CAP.TOKEN_PASS}")
-  end
-  c.filter_sensitive_data('private_access_token') do |interaction|
-    if interaction.request.uri.include? Settings.CAP.TOKEN_PATH
-      token_match = interaction.response.body.match(/"access_token":"(.*?)"/)
-      token_match.captures.first if token_match
-    end
-  end
-  c.filter_sensitive_data('CAP-LicenseID') do |interaction|
-    if interaction.request.uri.include? Settings.CAP.AUTHORSHIP_API_PATH
-      id_match = interaction.response.body.match(/"californiaPhysicianLicense".*?:.*?"(.*?)"/)
-      id_match.captures.first if id_match
-    end
-  end
-  c.filter_sensitive_data('CAP-UniversityID') do |interaction|
-    if interaction.request.uri.include? Settings.CAP.AUTHORSHIP_API_PATH
-      id_match = interaction.response.body.match(/"universityId".*?:.*?"(.*?)"/)
-      id_match.captures.first if id_match
-    end
-  end
-  c.filter_sensitive_data('CAP-UID') do |interaction|
-    if interaction.request.uri.include? Settings.CAP.AUTHORSHIP_API_PATH
-      id_match = interaction.response.body.match(/"uid".*?:.*?"(.*?)"/)
-      id_match.captures.first if id_match
-    end
-  end
-
-  # WOS API filters
-  c.filter_sensitive_data('Settings.WOS.AUTH_CODE') do |interaction|
-    Settings.WOS.AUTH_CODE if interaction.request.uri.include? 'WOKMWSAuthenticate'
-  end
-end
-
-def a_post(path)
-  a_request(:post, Settings.SCIENCEWIRE.BASE_URI + path)
-end
-
-def a_get(path)
-  a_request(:get, Settings.SCIENCEWIRE.BASE_URI + path)
-end
-
-def default_institution
-  Agent::AuthorInstitution.new(
-    Settings.HARVESTER.INSTITUTION.name,
-    Agent::AuthorAddress.new(Settings.HARVESTER.INSTITUTION.address.to_hash)
-  )
-end
-
-require_relative 'lib/web_of_science/wsdl'
-WebOfScience::WSDL.fetch
